Created basic files for MIKA in German
Small adaptations in config.yml (still a lot to do, choose models for tokenising etc.), some basic input provided in domain.yml, stories.yml and rules.yml, but there is still a lot of work to do on these. This pull request mainly focuses on the nlu.yml, but Mika should be up and running in a basic manner now. test_stories.yml is still very rudimentary at this stage.